// src/constants/enums.ts
export enum ErrorMessages {
    unknown = "Something went wrong!",
    invalid = "Invalid payload!",
    notFound = "User not found!",
    incorrectAuth = "Incorrect user or passowrd!",
    wrongPassword = "Passwords don't match!",
    duplicateMail = "Email already linked to another account!",
    incorrectId = "Invalid Id!"
}

export enum SuccessMessages {
    userCreated = "User was created successfully.",
    userDeleted = "User was successfully deleted."
}